requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.3.0
    # via
    #   -r requirements.in
    #   httpx
    #   starlette
    #   watchfiles
astroid==3.1.0
    # via
    #   -r requirements.in
    #   pylint
black==24.3.0
    # via -r requirements.in
build==1.2.1
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2024.2.2
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
cfgv==3.4.0
    # via
    #   -r requirements.in
    #   pre-commit
click==8.1.7
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
    #   uvicorn
dill==0.3.8
    # via
    #   -r requirements.in
    #   pylint
distlib==0.3.8
    # via
    #   -r requirements.in
    #   virtualenv
dnspython==2.6.1
    # via
    #   -r requirements.in
    #   email-validator
email-validator==2.1.1
    # via -r requirements.in
exceptiongroup==1.2.0
    # via
    #   -r requirements.in
    #   anyio
fastapi==0.110.0
    # via -r requirements.in
filelock==3.13.3
    # via
    #   -r requirements.in
    #   virtualenv
flake8==7.0.0
    # via -r requirements.in
greenlet==3.0.3
    # via
    #   -r requirements.in
    #   sqlalchemy
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via
    #   -r requirements.in
    #   httpx
httptools==0.6.1
    # via -r requirements.in
httpx==0.27.0
    # via -r requirements.in
identify==2.5.35
    # via
    #   -r requirements.in
    #   pre-commit
idna==3.6
    # via
    #   -r requirements.in
    #   anyio
    #   email-validator
    #   httpx
importlib-metadata==7.1.0
    # via
    #   -r requirements.in
    #   build
isort==5.13.2
    # via
    #   -r requirements.in
    #   pylint
itsdangerous==2.1.2
    # via -r requirements.in
jinja2==3.1.3
    # via -r requirements.in
markupsafe==2.1.5
    # via
    #   -r requirements.in
    #   jinja2
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
    #   pylint
mypy==1.9.0
    # via
    #   -r requirements.in
    #   sqlalchemy-stubs
mypy-extensions==1.0.0
    # via
    #   -r requirements.in
    #   black
    #   mypy
nodeenv==1.8.0
    # via
    #   -r requirements.in
    #   pre-commit
orjson==3.10.0
    # via -r requirements.in
packaging==24.0
    # via
    #   -r requirements.in
    #   black
    #   build
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.0
    # via
    #   -r requirements.in
    #   black
    #   pylint
    #   virtualenv
pre-commit==3.7.0
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pycodestyle==2.11.1
    # via
    #   -r requirements.in
    #   flake8
pydantic==2.6.4
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.16.3
    # via
    #   -r requirements.in
    #   pydantic
pydantic-extra-types==2.6.0
    # via -r requirements.in
pydantic-settings==2.2.1
    # via -r requirements.in
pyflakes==3.2.0
    # via
    #   -r requirements.in
    #   flake8
pylint==3.1.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-multipart==0.0.9
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   pre-commit
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
    #   httpx
sqlalchemy==2.0.29
    # via -r requirements.in
sqlalchemy-stubs==0.4
    # via -r requirements.in
starlette==0.36.3
    # via
    #   -r requirements.in
    #   fastapi
tomli==2.0.1
    # via
    #   -r requirements.in
    #   black
    #   build
    #   mypy
    #   pip-tools
    #   pylint
    #   pyproject-hooks
tomlkit==0.12.4
    # via
    #   -r requirements.in
    #   pylint
typing-extensions==4.10.0
    # via
    #   -r requirements.in
    #   anyio
    #   astroid
    #   black
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   pylint
    #   sqlalchemy
    #   sqlalchemy-stubs
    #   starlette
    #   uvicorn
ujson==5.9.0
    # via -r requirements.in
uvicorn==0.29.0
    # via -r requirements.in
uvloop==0.19.0
    # via -r requirements.in
virtualenv==20.25.1
    # via
    #   -r requirements.in
    #   pre-commit
watchfiles==0.21.0
    # via -r requirements.in
websockets==12.0
    # via -r requirements.in
wheel==0.43.0
    # via pip-tools
zipp==3.18.1
    # via
    #   -r requirements.in
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
